[norm] #26: five function per file: expansion/command_substitution.c.
