+++ includes/expansion/expansion.h
@@ -13,5 +13,6 @@ char				*parameter_expansion(char const *word);
 void				quote_removal(char **word);
 void				handle_quotes(char c, char *quoted);
 char				*tilde_expansion(char const *word);
+t_strlist			*split_subsitutions(char const *word);
 
 #endif
